+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- Add experimental visionOS support (#3328)
 - Add thread id and name to span data (#3359)
 
 ### Fixes